agent/requirements.txt
requests==2.32.3
pystray==0.19.5
pywin32==306; sys_platform == "win32"
pypiwin32==223; sys_platform == "win32"
Pillow==10.0.1
cairosvg==2.6.0
